refactored naming and fixed packaging
